+++ GbaGameEngine/src/engine/gba/graphics/tilemap/GBATilemapSet.cpp
@@ -16,7 +16,7 @@ TilemapSet::TilemapSet(
 	, const u32 * tileset
 	, const u32 tileSetDataCompressionFlags
 	, const u8 mapCount
-	, const u8 mapIsDynamicMask
+	, const u32 mapIsDynamicMask
 	, const u8 * mapTileWidths
 	, const u8 * mapTileHeights
 	, const u16 * mapData
@@ -32,7 +32,7 @@ TilemapSet::TilemapSet(
 
 	m_maps.Reserve(mapCount);
 
-	Bitmask<u8> mapIsDynamicBitMask = Bitmask<u8>(mapIsDynamicMask);
+	Bitmask<u32> mapIsDynamicBitMask = Bitmask<u32>(mapIsDynamicMask);
 
 	int mapDataOffset = 0;
 	for (u32 i = 0; i < mapCount; ++i)
